requirements.txt
pyautogui!=0.9.54
Pillow==9.5.0
pyscreeze==0.1.28
numpy==1.24.3
opencv-python==4.7.0.72
gymnasium==0.29.0
gymnasium[box2d]==0.29.0
tqdm==4.65.0
pytest-playwright==0.4.2
